Solved minor bugs on the scheduler and Settings. The function where the download_thread detects and selects the correct quality set by the user is now functioning.

+++ download_thread.py
@@ -4,6 +4,7 @@ import sys
 from wx import CallAfter
 import subprocess
 from pubsub import pub
+import json
 from threading import Thread
 import stopwatch
 import utilities as util
@@ -20,7 +21,7 @@ class Download(Thread):
         self.streamer = streamer
         self.url = streamer['url']
         self.name = streamer['name']
-        self.quality = streamer['quality']
+        self.userQuality = streamer['quality']
         self.dir = dir
 
         self.dl_total = 0
@@ -42,16 +43,20 @@ class Download(Thread):
         time_ended = time.strftime("%H:%M:%S")
         CallAfter(pub.sendMessage, topicName='log-stream-ended', streamer=self.name, time=time_ended)
 
-        # Changing the container of the stream to .mp4. This should be very fast.
-        ts = f"{self.dir}/{filename}.ts"
-        mp4 = f"{self.dir}/{filename}.mp4"
+        if 'audio' not in self.streamerQuality:
+            # Changing the container of the stream to .mp4. This should be very fast.
+            ts = f"{self.dir}/{filename}.ts"
+            mp4 = f"{self.dir}/{filename}.mp4"
 
-        subprocess.call(["ffmpeg", "-y", "-i", ts, "-vcodec", "copy", "-acodec", "copy", "-map", "0:v", "-map", "0:a", mp4], 
-            stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
-        os.remove(ts)
+            subprocess.call(["ffmpeg", "-y", "-i", ts, "-vcodec", "copy", "-acodec", "copy", "-map", "0:v", "-map", "0:a", mp4], 
+                stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
+            os.remove(ts)
+            
+            d = self.GetStreamerFromFile(self.name)
+            pub.sendMessage('add-to-queue', streamer=d)
 
-        pub.sendMessage('add-to-queue', streamer=self.streamer)
-        pub.sendMessage('remove-from-thread', name=self.name)
+            pub.sendMessage('remove-from-thread', name=self.name)
+
         sys.exit()
 
     def fetch_stream(self) -> bool:
@@ -60,8 +65,9 @@ class Download(Thread):
         # Will this catch streams end or stream offline? What about hostings? We don't want that.
         try:
             streams = self.session.streams(self.url)
-            print(streams.keys())
-            self.stream_data = streams['best'].open()
+            quality_list = list(streams.keys())
+            self.streamerQuality = self.ChooseQuality(self.userQuality, quality_list)
+            self.stream_data = streams[self.streamerQuality].open()
         except:
             return False
 
@@ -69,6 +75,8 @@ class Download(Thread):
 
     def start_download(self, filename: str):
         file = open(f"{self.dir}/{filename}.ts", "ab+")
+        CallAfter(pub.sendMessage, topicName='update-download-info', 
+            name=self.name, watch=None, quality=self.streamerQuality, size=None, speed=None)
  
         start = time.perf_counter()
         data = self.stream_data.read(1024)
@@ -82,7 +90,7 @@ class Download(Thread):
                 size, speed = util.get_progress_text(self.dl_total, self.dl_temp, diff)
                 start = time.perf_counter()
 
-                CallAfter(pub.sendMessage, topicName='update-download-info', name=self.name, watch=None, size=size, speed=speed)
+                CallAfter(pub.sendMessage, topicName='update-download-info', name=self.name, watch=None, quality=None, size=size, speed=speed)
                 self.dl_temp = 0
 
             file.write(data)
@@ -95,9 +103,89 @@ class Download(Thread):
 
         self.stopwatch.ping()
         CallAfter(pub.sendMessage, topicName='update-download-info', 
-        name=self.name, watch=self.stopwatch.to_str(), size=None, speed=None)
+        name=self.name, watch=self.stopwatch.to_str(), quality=None, size=None, speed=None)
 
     def KillDownloadThread(self):
         ''' Sets the `self.isActive` to False to end this thread. '''
 
         self.isActive = False
+
+
+    def ChooseQuality(self, quality: str, q_list: list) -> str:
+        ''' Given a `quality`, chooses the apropriate quality available in `q_list`.
+        Returns the exact quality chosen. '''
+        
+        q_list_sorted = []
+        for q in q_list:
+            if 'p' in q:
+                res = q.split('p')[0]
+                q_list_sorted.append(int(res))
+            else:
+                q_list_sorted.append(q)
+
+        index = -1
+
+        if quality == 'audio only':
+            for v in q_list_sorted:
+                if isinstance(v, str) and 'audio' in v:
+                    return v
+
+            quality = 'worst'
+
+        # Discoverin exactly what the best or worst are.
+        value = -1
+        if quality == 'best':
+            for i in range (0, len(q_list_sorted)):
+                if isinstance(q_list_sorted[i], int) and q_list_sorted[i] >= value:
+                    value = q_list_sorted[i]
+                    index = i
+
+            return q_list[index]
+
+        value = 9999
+        if quality == 'worst':
+            for i in range (0, len(q_list_sorted)):
+                if isinstance(q_list_sorted[i], int) and q_list_sorted[i] <= value:
+                    value = q_list_sorted[i]
+                    index = i
+
+            return q_list[index]
+
+        resolution = 0
+        match quality:
+            case 'high':
+                resolution = 720
+            case 'medium':
+                resolution = 480
+            case 'low':
+                resolution = 240
+            case _:
+                resolution = 1080
+
+        index = -1
+        for i in range(0, len(q_list_sorted)):
+            if isinstance(q_list_sorted[i], int):
+                value = q_list_sorted[i]
+            else:
+                continue
+
+            if value <= resolution:
+                index = i
+
+        return q_list[index]
+
+    def GetStreamerFromFile(self, name: str) -> dict:
+        ''' Search the app file, the dictionary from the streamer `name`. 
+        Getting this info through the file is necessary if the user edited the streamer
+        while he or she was live. '''
+
+        home = os.path.expanduser('~')
+        with open(f'{home}/.streamlink_looper.json', 'r', encoding='utf-8') as f:
+            text = f.read()
+            appData = json.loads(text)
+
+        for s in appData['streamers_data']:
+            if s['name'] == name:
+                return s
+
+        return None
